completions: Remove zsh completion files
Technically, we no longer need to install the common completion file now
as we are only using this file for bash completions. However, this issue
is a blocker for the upcoming platform-tools-31.0.2 release and I don't
want to spend a lot of time fiddling with completion files at this
point in time. The completion setup can still be cleanup up later.

Fixes #38