vendor: Don't use absolute dir for mkbootimg symlink
I noticed that the absolute symlink did not work on my system when
trying to execute the script inside it's installation directory before
the package is installed.
While investigating I noticed that on SUSE the symlink gets relinked
to be relative to be relative to the install location of the symlink
after which the link works.
Using a relative link seems safer to me.

Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>